### 2025-10-03 #9138 @editorialbot
[PRE REVIEW]: RegularizedOptimization.jl: A julia framework for regularized and nonsmooth optimization
**Submitting author:** @MaxenceGollier (<a href="http://orcid.org/0009-0008-3158-7912">Maxence Gollier</a>)
**Repository:** https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl/
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.2
