### 2025-10-03 #9138 @danielskatz
@MaxenceGollier - Please fix your .bib file - line 39 needs {}s around the value. This may be the case for line 50 and others as well.

Then use the command `@editorialbot check references` to check again. editorialbot commands need to be the first entry in a new comment.

This also might highlight missing or invalid DOIs that you can fix.

Additionally, your paper is too long. JOSS papers should be 250-1000 words, or close. The command `@editorialbot check repository` will check a number of things, including the paper's word count.

Then use the command `@editorialbot generate pdf` to make a new PDF.

Let me know when you think you are ready to proceed.
