+++ src/R2N.jl
@@ -1,6 +1,42 @@
 export R2N, R2NSolver, solve!
 
 import SolverCore.solve!
+using LinearAlgebra
+using LinearOperators
+
+# A small mutable wrapper that represents B + sigma*I without allocating a new
+# LinearOperator every time sigma or B changes. It provides mul! methods so it
+# can be used where a LinearOperator is expected.
+mutable struct ShiftedHessian{T}
+  B::Any
+  sigma::T
+end
+
+Base.size(op::ShiftedHessian) = size(op.B)
+Base.eltype(op::ShiftedHessian) = eltype(op.B)
+
+import LinearAlgebra: adjoint
+function adjoint(op::ShiftedHessian{T}) where T
+  return LinearAlgebra.Adjoint(op)
+end
+
+function LinearAlgebra.mul!(y::AbstractVector{T}, op::ShiftedHessian{T}, x::AbstractVector{T}) where T
+  mul!(y, op.B, x)
+  @inbounds for i in eachindex(y)
+    y[i] += op.sigma * x[i]
+  end
+  return y
+end
+
+function LinearAlgebra.mul!(y::AbstractVector{T}, opAd::Adjoint{<:Any,ShiftedHessian{T}}, x::AbstractVector{T}) where T
+  # Use the adjoint of the underlying operator and add sigma*x
+  mul!(y, adjoint(opAd.parent.B), x)
+  @inbounds for i in eachindex(y)
+    y[i] += opAd.parent.sigma * x[i]
+  end
+  return y
+end
+
 
 mutable struct R2NSolver{
   T <: Real,
@@ -32,6 +68,8 @@ mutable struct R2NSolver{
   Id::LinearOperator  # Identity operator
   x0_quad::V         # Zero vector for QuadraticModel x0
   reg_hess::LinearOperator  # regularized Hessian operator
+  reg_hess_wrapper::ShiftedHessian{T}  # mutable wrapper (B, sigma)
+  reg_hess_op::LinearOperator  # LinearOperator that captures the wrapper
 end
 
 function R2NSolver(
@@ -77,11 +115,20 @@ function R2NSolver(
   # So we need c = ∇fk, H = Bk + σI, c0 = 0
   σ = T(1)
   n = length(∇fk)
-  Id = opEye(T, n)  # Identity operator
-  x0_quad = zeros(T, n)  # Pre-allocate x0 for QuadraticModel
-  # Create QuadraticModel with H = Bk + σ*Id
-  reg_hess = Bk + σ * Id
-  sub_nlp = QuadraticModel(∇fk, reg_hess, c0 = zero(T), x0 = x0_quad, name = "R2N-subproblem")
+    Id = opEye(T, n)  # Identity operator
+    x0_quad = zeros(T, n)  # Pre-allocate x0 for QuadraticModel
+    # Create a mutable wrapper around the Hessian so we can update sigma/B without
+    # allocating a new operator every iteration.
+    reg_hess_wrapper = ShiftedHessian{T}(Bk, T(1))
+    # Create a LinearOperator that calls mul! on the wrapper. This operator is
+    # allocated once and keeps a reference to the mutable wrapper, so future
+    # updates can mutate the wrapper without reallocating the operator.
+  reg_hess_op = LinearOperator{T}(n, n, false, false,
+    (y, x) -> mul!(y, reg_hess_wrapper, x),
+    (y, x) -> mul!(y, adjoint(reg_hess_wrapper), x),
+    (y, x) -> mul!(y, adjoint(reg_hess_wrapper), x),
+  )
[…]
