### 2025-10-10 4a0d87c
Update src/LMTR_alg.jl
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
