### 2025-10-10 b06e017
Update src/R2N.jl
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
