+++ src/LM_alg.jl
@@ -29,6 +29,12 @@ mutable struct LMSolver{
   subsolver::ST
   subpb::PB
   substats::GenericExecutionStats{T, V, V, T}
+  # Preallocated QuadraticModel components to avoid reallocations
+  x0_quad::V
+  reg_hess_wrapper::ShiftedHessian{T}
+  reg_hess_op::LinearOperator
+  v0::V  # workspace for power method
+  tmp_res::V # temporary residual storage for power method
 end
 
 function LMSolver(
@@ -43,9 +49,10 @@ function LMSolver(
   xk = similar(x0)
   ∇fk = similar(x0)
   mν∇fk = similar(x0)
-  Fk = similar(x0, reg_nls.model.nls_meta.nequ)
-  Fkn = similar(Fk)
-  Jv = similar(Fk)
+  # residuals will be allocated after creating Jacobian operator (size may vary)
+  Fk = similar(x0, 0)
+  Fkn = similar(x0, 0)
+  Jv = similar(x0, 0)
   Jtv = similar(x0)
   xkn = similar(x0)
   s = similar(x0)
@@ -66,20 +73,40 @@ function LMSolver(
     has_bnds ? shifted(reg_nls.h, xk, l_bound_m_x, u_bound_m_x, reg_nls.selected) :
     shifted(reg_nls.h, xk)
 
-  Jk = jac_op_residual(reg_nls.model, xk)
-  # Compute initial residual
-  residual!(reg_nls.model, xk, Fk)
-  # Create LM subproblem using QuadraticModel: min s^T J^T F + 1/2 s^T (J^T J + σI) s
+  # Defer Jacobian/residual construction to solve!; preallocate JtF (length n)
   JtF = similar(xk)
-  mul!(JtF, Jk', Fk)
+  # residuals will be allocated in solve! when the Jacobian operator is available
+  Fk = similar(x0, 0)
+  Fkn = similar(x0, 0)
+  Jv = similar(x0, 0)
   n = length(xk)
-  Id = opEye(T, n)  # Identity operator
-  JtJ_plus_sigma = Jk' * Jk + T(1) * Id
   c0_val = dot(Fk, Fk) / 2  # Initial constant term = 1/2||F||²
-  sub_nlp = QuadraticModel(JtF, JtJ_plus_sigma, c0 = c0_val, x0 = zeros(T, n), name = "LM-subproblem")
+  x0_quad = zeros(T, n)  # Pre-allocated x0 for QuadraticModel
+  # Create mutable wrapper around Hessian; use JacobianGram to avoid forming J'J
+  gram = JacobianGram{T}(nothing, zeros(T, 0))
+  # Try to initialize JacobianGram.tmp with the row size of the Jacobian at x0
+  try
+    J_init = jac_op_residual(reg_nls.model, x0)
+    gram.J = J_init
+    gram.tmp = zeros(T, size(J_init, 1))
+  catch
+    # If jac_op_residual is not available for this model type, leave tmp empty
+    gram.J = nothing
+    gram.tmp = zeros(T, 0)
+  end
+  reg_hess_wrapper = ShiftedHessian{T}(gram, T(1))
+  reg_hess_op = LinearOperator{T}(n, n, false, false,
+    (y, x) -> mul!(y, reg_hess_wrapper, x),
+    (y, x) -> mul!(y, adjoint(reg_hess_wrapper), x),
+    (y, x) -> mul!(y, adjoint(reg_hess_wrapper), x),
+  )
+  sub_nlp = QuadraticModel(JtF, reg_hess_op, c0 = c0_val, x0 = x0_quad, name = "LM-subproblem")
[…]

[…]
