### 2025-10-11 8357983
Update src/utils.jl
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
