### 2025-10-11 589f75b
Update src/R2N.jl
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
