+++ src/LMTR_alg.jl
@@ -441,7 +441,7 @@ function SolverCore.solve!(
       # Update the QuadraticModel with new Jacobian and gradient
       JtF = Jk_update' * Fk
       JtJ = Jk_update' * Jk_update  
-      update_model!(solver.subpb.model, JtF, JtJ)
+      update_quadratic_model!(solver.subpb.model, JtF, JtJ)
     end
 
     if η2 ≤ ρk < Inf
