### 2025-10-12 2ea55a4
g   @warn "LM: subsolver iter count high" iter=solver.substats.iter fk=fk hk=hk σk=σk ν=ν   end   end   return (xk, fk, hk, σk, ν, solver) end
