+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @Article{         aravkin-baraldi-orban-2022,
   Author        = {A. Y. Aravkin and R. Baraldi and D. Orban},
   Title         = {A Proximal Quasi-{N}ewton Trust-Region Method for Nonsmooth Regularized Optimization},
-  Journal       = siopt,
+  Journal       = {SIAM J. Optim.},
   Year          = 2022,
   Volume        = 32,
   Number        = 2,
@@ -12,7 +12,7 @@
 @Article{         aravkin-baraldi-orban-2024,
   Author        = {A. Y. Aravkin and R. Baraldi and D. Orban},
   Title         = {A {L}evenberg–{M}arquardt Method for Nonsmooth Regularized Least Squares},
-  Journal       = sisc,
+  Journal       = {SIAM J. Sci. Comput.},
   Year          = 2024,
   Volume        = 46,
   Number        = 4,
@@ -23,7 +23,7 @@
 @Software{        leconte_linearoperators_jl_linear_operators_2023,
   Author        = {Leconte, Geoffroy and Orban, Dominique and Soares Siqueira, Abel and contributors},
   license       = {MPL-2.0},
-  Title         = {{LinearOperators.jl: Linear Operators for Julia}},
+  Title         = {{LinearOperators.jl}: Linear Operators for Julia},
   url           = {https://github.com/JuliaSmoothOptimizers/LinearOperators.jl},
   version       = {2.6.0},
   Year          = 2023,
@@ -32,34 +32,33 @@
 @Article{         leconte-orban-2023,
   Author        = {G. Leconte and D. Orban},
   Title         = {The Indefinite Proximal Gradient Method},
-  Journal       = coap,
+  Journal       = {Comput. Optim. Appl.},
   Year          = 2025,
   Volume        = 91,
   Number        = 2,
-  Pages         = 861--903,
+  Pages         = {861--903},
   doi           = {10.1007/s10589-024-00604-5},
 }
 
 @TechReport{      leconte-orban-2023-2,
   Author        = {G. Leconte and D. Orban},
   Title         = {Complexity of trust-region methods with unbounded {H}essian approximations for smooth and nonsmooth optimization},
-  Institution   = gerad,
+  Institution   = {GERAD},
   Year          = 2023,
   Type          = {Cahier},
   Number        = {G-2023-65},
-  Address       = gerad-address,
+  Address       = {Montr\'eal, QC, Canada},
   url           = {https://www.gerad.ca/fr/papers/G-2023-65},
 }
 
 @TechReport{      diouane-habiboullah-orban-2024,
   Author        = {Youssef Diouane and Mohamed Laghdaf Habiboullah and Dominique Orban},
-  Title         = {A proximal modified quasi-Newton method for nonsmooth regularized optimization},
+  Title         = {A proximal modified quasi-{N}ewton method for nonsmooth regularized optimization},
   Institution   = {GERAD},
   Year          = 2024,
   Type          = {Cahier},
   Number        = {G-2024-64},
   Address       = {Montr\'eal, Canada},
-  doi           = {10.48550/arxiv.2409.19428},
   url           = {https://www.gerad.ca/fr/papers/G-2024-64},
 }
 
@@ -74,47 +73,47 @@
   doi           = {10.13140/RG.2.2.16095.47527},
 }
 
-@Misc{orban-siqueira-cutest-2020,
-  author = {D. Orban and A. S. Siqueira and {contributors}},
-  title = {{CUTEst.jl}: {J}ulia's {CUTEst} interface},
-  month = {October},
-  url = {https://github.com/JuliaSmoothOptimizers/CUTEst.jl},
-  year = {2020},
-  DOI = {10.5281/zenodo.1188851},
[…]
