### 2025-10-15 a434feb
Update callback behavior in solvers to return false on user exit request
