+++ src/R2DH.jl
@@ -213,7 +213,7 @@ function SolverCore.solve!(
   solver::R2DHSolver{T},
   reg_nlp::AbstractRegularizedNLPModel{T, V},
   stats::GenericExecutionStats{T, V};
-  callback = (args...) -> nothing,
+  callback = (args...) -> false,
   x::V = reg_nlp.model.meta.x0,
   atol::T = √eps(T),
   rtol::T = √eps(T),
@@ -338,12 +338,15 @@ function SolverCore.solve!(
     error("R2DH: prox-gradient step should produce a decrease but ξ = $(ξ)")
   atol += rtol * sqrt_ξ_νInv # make stopping test absolute and relative
 
+  user_requested_exit = callback(nlp, solver, stats) :: Bool
+
   set_status!(
     stats,
     get_status(
       reg_nlp,
       elapsed_time = stats.elapsed_time,
       iter = stats.iter,
+      user_requested_exit = user_requested_exit,
       optimal = solved,
       improper = improper,
       max_eval = max_eval,
@@ -351,8 +354,6 @@ function SolverCore.solve!(
       max_iter = max_iter,
     ),
   )
-
-  callback(nlp, solver, stats)
 
   done = stats.status != :unknown
 
@@ -433,12 +434,15 @@ function SolverCore.solve!(
     (ξ < 0 && sqrt_ξ_νInv > neg_tol) &&
       error("R2DH: prox-gradient step should produce a decrease but ξ = $(ξ)")
 
+    user_requested_exit = callback(nlp, solver, stats) :: Bool
+
     set_status!(
       stats,
       get_status(
         reg_nlp,
         elapsed_time = stats.elapsed_time,
         iter = stats.iter,
+        user_requested_exit = user_requested_exit,
         optimal = solved,
         improper = improper,
         max_eval = max_eval,
@@ -446,8 +450,6 @@ function SolverCore.solve!(
         max_iter = max_iter,
       ),
     )
-
-    callback(nlp, solver, stats)
 
     done = stats.status != :unknown
   end
