+++ src/TR_alg.jl
@@ -198,7 +198,7 @@ function SolverCore.solve!(
   solver::TRSolver{T, G, V},
   reg_nlp::AbstractRegularizedNLPModel{T, V},
   stats::GenericExecutionStats{T, V};
-  callback = (args...) -> nothing,
+  callback = (args...) -> false,
   x::V = reg_nlp.model.meta.x0,
   atol::T = √eps(T),
   sub_atol::T = √eps(T),
@@ -340,12 +340,15 @@ function SolverCore.solve!(
   atol += rtol * sqrt_ξ1_νInv # make stopping test absolute and relative
   sub_atol += rtol * sqrt_ξ1_νInv
 
+  user_requested_exit = callback(nlp, solver, stats) :: Bool
+
   set_status!(
     stats,
     get_status(
       reg_nlp,
       elapsed_time = stats.elapsed_time,
       iter = stats.iter,
+      user_requested_exit = user_requested_exit,
       optimal = solved,
       improper = improper,
       max_eval = max_eval,
@@ -353,8 +356,6 @@ function SolverCore.solve!(
       max_iter = max_iter,
     ),
   )
-
-  callback(nlp, solver, stats)
 
   done = stats.status != :unknown
   while !done
@@ -499,12 +500,15 @@ function SolverCore.solve!(
     (ξ1 < 0 && sqrt_ξ1_νInv > neg_tol) &&
       error("TR: prox-gradient step should produce a decrease but ξ1 = $(ξ1)")
 
+    user_requested_exit = callback(nlp, solver, stats) :: Bool
+
     set_status!(
       stats,
       get_status(
         reg_nlp,
         elapsed_time = stats.elapsed_time,
         iter = stats.iter,
+        user_requested_exit = user_requested_exit,
         optimal = solved,
         improper = improper,
         max_eval = max_eval,
@@ -512,8 +516,6 @@ function SolverCore.solve!(
         max_iter = max_iter,
       ),
     )
-
-    callback(nlp, solver, stats)
 
     done = stats.status != :unknown
   end
