+++ src/R2N.jl
@@ -340,7 +340,6 @@ function SolverCore.solve!(
     error("R2N: prox-gradient step should produce a decrease but ξ1 = $(ξ1)")
   atol += rtol * sqrt_ξ1_νInv # make stopping test absolute and relative
 
-
   user_requested_exit = callback(nlp, solver, stats) :: Bool
 
   set_status!(
@@ -488,7 +487,6 @@ function SolverCore.solve!(
 
     (ξ1 < 0 && sqrt_ξ1_νInv > neg_tol) &&
       error("R2N: prox-gradient step should produce a decrease but ξ1 = $(ξ1)")
-    
     user_requested_exit = callback(nlp, solver, stats) :: Bool
 
     set_status!(
