### 2025-10-15 734bc23
Add sigma_cauchy and scp_norm to solver statistics in LMTR, LMSolver, R2DH, R2N, and R2Solver
