+++ src/TRDH_alg.jl
@@ -333,6 +333,7 @@ function SolverCore.solve!(
   set_objective!(stats, fk + hk)
   set_solver_specific!(stats, :smooth_obj, fk)
   set_solver_specific!(stats, :nonsmooth_obj, hk)
+  set_solver_specific!(stats, :sigma_cauchy, 1/ν)
 
   # models
   φ1 = let ∇fk = ∇fk
@@ -358,6 +359,7 @@ function SolverCore.solve!(
 
   if reduce_TR
     prox!(s, ψ, mν∇fk, ν)
+    set_solver_specific!(stats, :scp_norm, norm(s))
     mks = mk1(s)
 
     ξ1 = hk - mks + max(1, abs(hk)) * 10 * eps()
@@ -383,6 +385,7 @@ function SolverCore.solve!(
   sNorm = χ(s)
 
   if !reduce_TR
+    set_solver_specific!(stats, :scp_norm, norm(s))
     sqrt_ξ_νInv = ξ ≥ 0 ? sqrt(ξ / ν) : sqrt(-ξ / ν)
     solved = (ξ < 0 && sqrt_ξ_νInv ≤ neg_tol) || (ξ ≥ 0 && sqrt_ξ_νInv < atol)
     (ξ < 0 && sqrt_ξ_νInv > neg_tol) &&
@@ -473,10 +476,12 @@ function SolverCore.solve!(
     set_time!(stats, time() - start_time)
 
     ν = reduce_TR ? (α * Δk)/(DNorm + one(T)) : α / (DNorm + one(T))
+    set_solver_specific!(stats, :sigma_cauchy, 1/ν)
     mν∇fk .= -ν .* ∇fk
 
     if reduce_TR
       prox!(s, ψ, mν∇fk, ν)
+      set_solver_specific!(stats, :scp_norm, norm(s))
       ξ1 = hk - mk1(s) + max(1, abs(hk)) * 10 * eps()
       sqrt_ξ_νInv = ξ1 ≥ 0 ? sqrt(ξ1 / ν) : sqrt(-ξ1 / ν)
       solved = (ξ1 < 0 && sqrt_ξ_νInv ≤ neg_tol) || (ξ1 ≥ 0 && sqrt_ξ_νInv < atol)
@@ -489,6 +494,7 @@ function SolverCore.solve!(
     sNorm = χ(s)
 
     if !reduce_TR
+      set_solver_specific!(stats, :scp_norm, norm(s))
       ξ = hk - mk(s) + max(1, abs(hk)) * 10 * eps()
       sqrt_ξ_νInv = ξ ≥ 0 ? sqrt(ξ / ν) : sqrt(-ξ / ν)
       solved = (ξ < 0 && sqrt_ξ_νInv ≤ neg_tol) || (ξ ≥ 0 && sqrt_ξ_νInv < atol)
