### 2025-10-16 974cc47
paper: reduce number of words
