### 2025-10-16 ce435f5
paper: further reduce word count
