### 2025-10-16 f52a582
paper: further reduce word count
