+++ paper/examples/Benchmark.tex
@@ -5,9 +5,4 @@
   R2N (LSR1, SVM) & first\_order & 1.9511 & 185 & 101 & 27932 & 192.493 \\
   LM (SVM) & first\_order & 19.7826 & 6 & 2876 & 1001 & 201.186 \\
   LMTR (SVM) & first\_order & 12.4967 & 11 & 1614 & 432 & 188.274 \\
-  \hline
-  TR (LBFGS, NNMF) & first\_order & 0.1014 & 42 & 40 & 3160 & 976.06 \\
-  R2N (LBFGS, NNMF) & first\_order & 0.4913 & 169 & 107 & 17789 & 411.727 \\
-  LM (NNMF) & first\_order & 0.1157 & 14 & 7042 & 2601 & 131.184 \\
-  LMTR (NNMF) & first\_order & 0.0697 & 9 & 4066 & 1435 & 131.186 \\\hline
 \end{tabular}
