### 2025-10-16 cf63b11
paper: rename "Examples" section to "Example"
