### 2025-10-16 e551d8b
paper: apply corrections from Mohamed
