+++ paper/paper.bib
@@ -101,7 +101,7 @@
 
 @Software{       migot-orban-siqueira-optimizationproblems-2023,
   author       = {T. Migot and D. Orban and A. S. Siqueira},
-  title        = {{OptimizationProblems.jl}: A collection of optimization problems in Julia},
+  title        = {{OptimizationProblems.jl}: A collection of optimization problems in {J}ulia},
   year         = {2023},
   doi          = {10.5281/zenodo.3672094},
   url          = {https://github.com/JuliaSmoothOptimizers/OptimizationProblems.jl},
@@ -139,7 +139,7 @@
 
 @Article{         themelis-stella-patrinos-2017,
   Author        = {Themelis, Andreas and Stella, Lorenzo and Patrinos, Panagiotis},
-  Title         = {Forward-Backward Envelope for the Sum of Two Nonconvex Functions: Further Properties and Nonmonotone line seach Algorithms},
+  Title         = {Forward-Backward Envelope for the Sum of Two Nonconvex Functions: Further Properties and Nonmonotone line search Algorithms},
   Journal       = {SIAM J. Optim.},
   Year          = 2018,
   Volume        = 28,
