### 2025-10-16 195c809
paper: remove non-monotone
