+++ src/LM_alg.jl
@@ -149,7 +149,7 @@ lower semi-continuous, proper and prox-bounded.
 
 At each iteration, a step s is computed as an approximate solution of
 
-    min  ½ ‖J(x) s + F(x)‖² + ½ σ ‖s‖² + ψ(s; x)
+    min  ½ ‖J(x) s + F(x)‖² + ½ σ ‖s‖² + ψ(s; xₖ) 
 
 where F(x) and J(x) are the residual and its Jacobian at x, respectively, ψ(s; xₖ) is either h(xₖ + s) or an approximation of h(xₖ + s),
 ‖⋅‖ is the ℓ₂ norm and σₖ > 0 is the regularization parameter.
@@ -196,6 +196,8 @@ function LM(nls::AbstractNLSModel, h::H, options::ROSolverOptions; kwargs...) wh
   kwargs_dict = Dict(kwargs...)
   selected = pop!(kwargs_dict, :selected, 1:(nls.meta.nvar))
   x0 = pop!(kwargs_dict, :x0, nls.meta.x0)
+  # allow callers to request skipping σ in the quadratic regularizer
+  skip_sigma = pop!(kwargs_dict, :skip_sigma, false)
   reg_nls = RegularizedNLPModel(nls, h, selected)
   return LM(
     reg_nls;
@@ -211,6 +213,7 @@ function LM(nls::AbstractNLSModel, h::H, options::ROSolverOptions; kwargs...) wh
     η1 = options.η1,
     η2 = options.η2,
     γ = options.γ,
+    skip_sigma = skip_sigma,
     kwargs_dict...,
   )
 end
@@ -219,9 +222,11 @@ function LM(reg_nls::AbstractRegularizedNLPModel; kwargs...)
   kwargs_dict = Dict(kwargs...)
   subsolver = pop!(kwargs_dict, :subsolver, R2Solver)
   m_monotone = pop!(kwargs_dict, :m_monotone, 1)
+  # propagate skip_sigma to the solver
+  skip_sigma = pop!(kwargs_dict, :skip_sigma, false)
   solver = LMSolver(reg_nls, subsolver = subsolver, m_monotone = m_monotone)
   stats = RegularizedExecutionStats(reg_nls)
-  solve!(solver, reg_nls, stats; kwargs_dict...)
+  solve!(solver, reg_nls, stats; kwargs_dict..., skip_sigma = skip_sigma)
   return stats
 end
 
@@ -245,6 +250,7 @@ function SolverCore.solve!(
   η2::T = T(0.9),
   γ::T = T(3),
   θ::T = 1/(1 + eps(T)^(1 / 5)),
+  skip_sigma::Bool = false,   # <-- new kwarg (default: false)
 ) where {T, V, G}
   reset!(stats)
 
@@ -313,6 +319,8 @@ function SolverCore.solve!(
   local ξ1::T
   local ρk::T = zero(T)
   local prox_evals::Int = 0
+  local inner_stall_count::Int = 0
+  INNER_STALL_THRESHOLD = 500
   # Track subsolver stalls and keep a backup step
   stall_counter = 0
   s_backup = similar(s)
@@ -339,7 +347,12 @@ function SolverCore.solve!(
     solver.tmp_res = zeros(T, size(Jk, 1))
   end
   σmax = power_method_singular!(Jk, solver.v0, solver.subpb.model.data.v, solver.tmp_res, 5)
-  ν = θ / (σmax^2 + σk) # ‖J'J + σₖ I‖ = ‖J‖² + σₖ
+  # if the caller requested skipping σ, compute ν without adding σk into the denominator
+  if skip_sigma
+    ν = θ / (σmax^2)
[…]
