### 2025-10-17 109c004
Update the in-loop status update to recompute the stationarity measure and solved flag from the current ξ and ν before calling set_status!.
