### 2025-10-17 ef6870c
paper: Apply suggestions from dpo review
Co-authored-by: Dominique <dominique.orban@gmail.com>
