### 2025-10-17 39bde6d
paper: add workflow to count number of words
