### 2025-10-17 6287508
fix wordcount workflow
