+++ .github/workflows/wordcount.yml
@@ -4,6 +4,11 @@ on:
   pull_request:
     types: [opened, synchronize, reopened]
 
+permissions:
+  contents: read
+  pull-requests: write
+  issues: write
+
 jobs:
   count-words:
     runs-on: ubuntu-latest
@@ -18,7 +23,7 @@ jobs:
       - name: Count words in paper.md
         id: wordcount
         run: |
-          COUNT=$(pandoc paper.md -t plain | wc -w)
+          COUNT=$(pandoc paper/paper.md -t plain | wc -w)
           echo "count=$COUNT" >> $GITHUB_OUTPUT
 
       - name: Comment on PR with word count
