### 2025-10-17 269d7bc
comment wordcount only if from same repo
