### 2025-10-17 ca5113c
paper: apply suggestions from D1Lab
