+++ paper/examples/Benchmark.tex
@@ -1,8 +1,8 @@
 \begin{tabular}{lcrrrrr}
   \hline
   \textbf{Method} & \textbf{Status} & \textbf{$t$($s$)} & \textbf{$\#f$} & \textbf{$\#\nabla f$} & \textbf{$\#prox$} & \textbf{Objective} \\\hline
-  TR (LSR1, SVM) & first\_order & 3.9349 & 347 & 291 & 4037 & 179.837 \\
-  R2N (LSR1, SVM) & first\_order & 1.9511 & 185 & 101 & 27932 & 192.493 \\
-  LM (SVM) & first\_order & 19.7826 & 6 & 2876 & 1001 & 201.186 \\
-  LMTR (SVM) & first\_order & 12.4967 & 11 & 1614 & 432 & 188.274 \\
+  TR (LSR1) & first\_order & 3.9349 & 347 & 291 & 4037 & 179.837 \\
+  R2N (LSR1) & first\_order & 1.9511 & 185 & 101 & 27932 & 192.493 \\
+  LM & first\_order & 19.7826 & 6 & 2876 & 1001 & 201.186 \\
+  LMTR & first\_order & 12.4967 & 11 & 1614 & 432 & 188.274 \\
 \end{tabular}
