### 2025-10-17 17f2130
paper: remove more words
