### 2025-10-17 fffd47c
paper: remove more words based on the suggestions of D1Lab
