### 2025-10-17 deef93e
Update src/LMTR_alg.jl
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
