### 2025-10-17 fb0fbb3
Update src/R2N.jl
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
