### 2025-10-17 8dd1b2a
Copilot remarks
