### 2025-10-17 e176915
paper: remove more words, remove tilde
