### 2025-10-17 f5d105e
solver error
