### 2025-10-17 d9ff128
Apply suggestions from code review
Co-authored-by: Dominique <dominique.orban@gmail.com>
