### 2025-10-19 94712ce
remove LSR1 from benchmark table
