### 2025-10-20 c161c84
numerical results: apply D1Lab comments
