### 2025-10-20 7d40b0f
revert some changes in the paper
