### 2025-10-20 75dd44e
Remove NNMF problem and update Benchmark.jl accordingly
