+++ paper/examples/Benchmark.jl
@@ -3,7 +3,6 @@
 # ======== IMPORTS ======== #
 #############################
 using Random, LinearAlgebra
-using ProximalOperators, ProximalCore, ProximalAlgorithms
 using ShiftedProximalOperators
 using NLPModels, NLPModelsModifiers
 using RegularizedOptimization, RegularizedProblems
@@ -14,7 +13,7 @@ using LaTeXStrings
 
 # Local includes
 include("comparison-config.jl")
-using .ComparisonConfig: CFG, CFG2
+using .ComparisonConfig: CFG
 
 #############################
 # ===== Helper utils ====== #
@@ -58,7 +57,7 @@ function run_tr_svm!(model, x0; λ = 1.0, qn = :LSR1, atol = 1e-3, rtol = 1e-3, 
     t = @elapsed RegularizedOptimization.solve!(solver, reg_nlp, stats;
         x = x0, atol = atol, rtol = rtol, verbose = verbose, sub_kwargs = sub_kwargs)
     return (
-        name      = "TR ($(String(qn)), SVM)",
+        name      = "TR",
         status    = string(stats.status),
         time      = t,
         iters     = get(stats.solver_specific, :outer_iter, missing),
@@ -84,7 +83,7 @@ function run_r2n_svm!(model, x0; λ = 1.0, qn = :LBFGS, atol = 1e-3, rtol = 1e-3
     t = @elapsed RegularizedOptimization.solve!(solver, reg_nlp, stats;
         x = x0, atol = atol, rtol = rtol, verbose = verbose, sub_kwargs = sub_kwargs)
     return (
-        name      = "R2N ($(String(qn)), SVM)",
+        name      = "R2N",
         status    = string(stats.status),
         time      = t,
         iters     = get(stats.solver_specific, :outer_iter, missing),
@@ -108,7 +107,7 @@ function run_LM_svm!(nls_model, x0; λ = 1.0, atol = 1e-3, rtol = 1e-3, verbose 
     t = @elapsed RegularizedOptimization.solve!(solver, reg_nls, stats;
         x = x0, atol = atol, rtol = rtol, verbose = verbose, sub_kwargs = sub_kwargs)
     return (
-        name      = "LM (SVM)",
+        name      = "LM",
         status    = string(stats.status),
         time      = t,
         iters     = get(stats.solver_specific, :outer_iter, missing),
@@ -132,7 +131,7 @@ function run_LMTR_svm!(nls_model, x0; λ = 1.0, atol = 1e-3, rtol = 1e-3, verbos
     t = @elapsed RegularizedOptimization.solve!(solver, reg_nls, stats;
         x = x0, atol = atol, rtol = rtol, verbose = verbose, sub_kwargs = sub_kwargs)
     return (
-        name      = "LMTR (SVM)",
+        name      = "LMTR",
         status    = string(stats.status),
         time      = t,
         iters     = get(stats.solver_specific, :outer_iter, missing),
@@ -185,178 +184,23 @@ function bench_svm!(cfg = CFG)
     return data_svm
 end
 
-#############################
-# ======= NNMF bench ====== #
-#############################
-
-function run_tr_nnmf!(model, x0; λ = 1.0, qn = :LSR1, atol = 1e-3, rtol = 1e-3, verbose = 0, sub_kwargs = (;), selected = nothing)
-    qn_model = ensure_qn(model, qn)
-    reset!(qn_model)
-    reg_nlp  = RegularizedNLPModel(qn_model, NormL0(λ), selected)
-    solver   = TRSolver(reg_nlp)
-    stats    = RegularizedExecutionStats(reg_nlp)
[…]
