### 2025-10-21 be59180
Add scp_norm to solver-specific entries in RegularizedExecutionStats
