+++ src/utils.jl
@@ -128,6 +128,7 @@ function RegularizedExecutionStats(reg_nlp::AbstractRegularizedNLPModel{T, V}) w
   set_solver_specific!(stats, :sigma_cauchy, T(Inf))
   set_solver_specific!(stats, :radius, T(Inf))
   set_solver_specific!(stats, :prox_evals, T(Inf))
+  set_solver_specific!(stats, :scp_norm, T(Inf))
   return stats
 end
 
