### 2025-10-21 fe6762f
Refactor SVM benchmark output formatting and improve LaTeX table generation
