### 2025-10-21 cf96428
Apply suggestions from code review
Co-authored-by: Dominique <dominique.orban@gmail.com>
