### 2025-10-21 83d9bc8
Add DOI to paper.bib entry
