+++ paper/paper.bib
@@ -156,5 +156,6 @@
   number        = {1},
   pages         = {293--318},
   year          = {1992},
-  publisher     = {Springer}
+  publisher     = {Springer},
+  doi           = {10.1007/BF01581204}
 }
