### 2025-10-21 badc31f
Remove NNMF problem and update Benchmark.jl accordingly (#254)
* Remove NNMF problem and update Benchmark.jl accordingly

* Refactor SVM benchmark output formatting and improve LaTeX table generation
