### 2025-11-03 #9344 @editorialbot
[REVIEW]: RegularizedOptimization.jl: A julia framework for regularized and nonsmooth optimization
**Submitting author:** @MaxenceGollier (<a href="http://orcid.org/0009-0008-3158-7912">Maxence Gollier</a>)
**Repository:** https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl/
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
