+++ src/iR2N.jl
@@ -347,9 +347,18 @@ function SolverCore.solve!(
   prox!(s1, ψ, mν∇fk, ν₁)
   mks = mk1(s1)
 
+  # Record outer-level prox call counters (exclude feasibility prox above)
+  if ψ isa InexactShiftedProximableFunction
+    ctx = ψ.h.context
+    ctx.prox_calls_outer += 1
+    ctx.prox_iters_outer += ctx.last_prox_iters
+  end
+
   ξ1 = hk - mks + max(1, abs(hk)) * 10 * eps()
   sqrt_ξ1_νInv = ξ1 ≥ 0 ? sqrt(ξ1 / ν₁) : sqrt(-ξ1 / ν₁)
-  solved = (ξ1 < 0 && sqrt_ξ1_νInv ≤ neg_tol) || (ξ1 ≥ 0 && sqrt_ξ1_νInv ≤ atol * √κξ)
+  solved =
+    (ξ1 < 0 && sqrt_ξ1_νInv ≤ neg_tol && stats.iter > 10) ||
+    (ξ1 ≥ 0 && sqrt_ξ1_νInv ≤ atol && stats.iter > 10)# * √κξ)
   (ξ1 < 0 && sqrt_ξ1_νInv > neg_tol) &&
     error("iR2N: prox-gradient step should produce a decrease but ξ1 = $(ξ1)")
   atol += rtol * sqrt_ξ1_νInv # make stopping test absolute and relative
@@ -378,6 +387,12 @@ function SolverCore.solve!(
     solver.subpb.model.σ = σk
     isa(solver.subsolver, R2DHSolver) && (solver.subsolver.D.d[1] = 1 / ν₁)
     ν_sub = isa(solver.subsolver, R2DHSolver) ? 1 / σk : ν₁
+    # Snapshot inner context counters to aggregate deltas only
+    inner_ctx = solver.subpb.h.h.context
+    prev_calls_inner = inner_ctx.prox_calls_inner
+    prev_iters_inner = inner_ctx.prox_iters_inner
+    prev_proxstats3_inner = inner_ctx.prox_stats[3]
+
     solve!(
       solver.subsolver,
       solver.subpb,
@@ -392,8 +407,17 @@ function SolverCore.solve!(
     s .= solver.substats.solution
 
     if ψ isa InexactShiftedProximableFunction
-      ψ.h.context.prox_stats[2] += solver.substats.iter
-      ψ.h.context.prox_stats[3] += solver.substats.solver_specific[:ItersProx]
+      # Aggregate inner-level prox counters by delta since last call
+      calls_delta = inner_ctx.prox_calls_inner - prev_calls_inner
+      iters_delta = inner_ctx.prox_iters_inner - prev_iters_inner
+      iters_prox_delta = inner_ctx.prox_stats[3] - prev_proxstats3_inner
+      let outer_ctx = ψ.h.context
+        outer_ctx.prox_calls_inner += calls_delta
+        outer_ctx.prox_iters_inner += iters_delta
+        outer_ctx.prox_stats[3] += iters_prox_delta
+        # Keep subsolver iteration total for legacy metric
+        outer_ctx.prox_stats[2] += solver.substats.iter
+      end
     end
 
     if norm(s) > β * norm(s1)
@@ -415,13 +439,10 @@ function SolverCore.solve!(
       error("iR2N: failed to compute a step: ξ = $ξ and sqrt_ξ_νInv = $sqrt_ξ_νInv")
     end
 
-    aux_assert = (1 - 1 / (1 + θ)) * ξ1
+    aux_assert = (1 - θ) * ξ1
     if ξ < aux_assert && sqrt_ξ_νInv > neg_tol
-      @warn(
[…]
