### 2025-11-28 8d02ac3
bump version to 0.2.0
