### 2025-11-30 807aa4c
revert to v0.1.0 and update author list
