### 2025-11-30 3f78b18
Fix author list formatting
