+++ Project.toml
@@ -1,7 +1,7 @@
 name = "RegularizedOptimization"
 uuid = "196f2941-2d58-45ba-9f13-43a2532b2fa8"
 author = ["Robert Baraldi <rbaraldi@uw.edu>", 
-          "Youssef Diouane <youssef.diouane@polymtl.ca>"
+          "Youssef Diouane <youssef.diouane@polymtl.ca>",
           "Maxence Gollier <maxence.gollier@polymtl.ca>",
           "Mohamed Laghdaf Habiboullah <mohamed.habiboullah@polymtl.ca>",
           "Dominique Orban <dominique.orban@gmail.com>"]
