### 2025-11-30 2f8c5a4
Add Geoffroy to author list
