### 2025-12-01 JuliaSmoothOptimizers/RegularizedOptimization.jl#259 @pierre-borie
JOSS Review: Functionality (paper example)
- [ ] Functionality: the example given in the paper did not work for me with the packages listed. In particular, it required to install
   - `NLPModelsModifiers.jl` to define the `LSR1Model`
   - `ShiftedProximalOperators.jl` to define the `RootNormLhalf` regularizer

I tried running the example with Julia 1.12.2, in case the example works on a earlier version.

Another minor question / remark: is the package planed to be registered? Installation via the github repostory is not an issue per se but most of the JSO packages seem to be on JuliaRegisteries.

https://github.com/openjournals/joss-reviews/issues/9344
