### 2025-12-02 5eaf39f
Correct author email addresses in Project.toml
Updated author email addresses for Maxence Gollier and Mohamed Laghdaf Habiboullah.
