### 2025-12-03 a31a8c2
iR2N: R2N for inexact evaluations
