### 2025-12-03 558b995
iR2N: R2N for inexact evaluations
