+++ test/test_allocs.jl
@@ -59,6 +59,20 @@ end
           @test @wrappedallocs(solve!(solver, reg_nlp, stats, atol = 1e-6, rtol = 1e-6)) == 0
         @test stats.status == :first_order
       end
+      # TODO update this test
+      # for (h, h_name) ∈ ((NormL1(0.1), "l1"), (NormLp(0.1, 1.6, bpdn.meta.nvar), "lp"))
+      #   for (solver, solver_name) ∈ ((:iR2Solver, "iR2"),)
+      #     @testset "$(solver_name)" begin
+      #       reg_nlp = RegularizedNLPModel(LBFGSModel(bpdn), h)
+      #       solver = eval(solver)(reg_nlp)
+      #       stats = RegularizedExecutionStats(reg_nlp)
+      #       @test @wrappedallocs(
+      #         solve!(solver, reg_nlp, stats, ν = 1.0, atol = 1e-6, rtol = 1e-6, verbose = 0)
+      #       ) == 0
+      #       @test stats.status == :first_order
+      #     end
+      #   end
+      # end
     end
     @testset "Augmented Lagrangian" begin
       continue # FIXME : fails due to type instabilities in ADNLPModels...
