+++ previews/PR260/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"examples/basic/#A-regularized-optimization-problem","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"In this tutorial, we will show how to model and solve the nonconvex nonsmooth optimization problem\n\n  min_x in mathbbR^2 (1 - x_1)^2 + 100(x_2 - x_1^2)^2 + x_1 + x_2\n\nwhich can be seen as a ell_1 regularization of the Rosenbrock function.  It can be shown that the solution to the problem is \n\n  x^* = beginpmatrix\n  025\n  00575\n  endpmatrix","category":"section"},{"location":"examples/basic/#Modelling-the-problem","page":"A regularized optimization problem","title":"Modelling the problem","text":"We first formulate the objective function as the sum of a smooth function f and a nonsmooth regularizer h:\n\n  (1 - x_1)^2 + 100(x_2 - x_1^2)^2 + x_1 + x_2 = f(x_1 x_2) + h(x_1 x_2)\n\nwhere \n\nbeginalign*\nf(x_1 x_2) = (1 - x_1)^2 + 100(x_2 - x_1^2)^2\nh(x_1 x_2) = x_1\nendalign*\n\nTo model f, we are going to use ADNLPModels.jl. For the nonsmooth regularizer, we use ProximalOperators.jl.  We then wrap the smooth function and the regularizer in a RegularizedNLPModel\n\nusing ADNLPModels\nusing ProximalOperators\nusing RegularizedProblems\n\n# Model the function\nf_fun = x -> (1 - x[1])^2 + 100*(x[2] - x[1]^2)^2\n\n# Choose a starting point for the optimization process, for the sake of this example, we choose\nx0 = [-1.0, 2.0]\n\n# Get an NLPModel corresponding to the smooth function f\nf_model = ADNLPModel(f_fun, x0, name = \"AD model of f\") \n\n# Get the regularizer from ProximalOperators\nh = NormL1(1.0)\n\n# Wrap into a RegularizedNLPModel\nregularized_pb = RegularizedNLPModel(f_model, h)","category":"section"},{"location":"examples/basic/#Solving-the-problem","page":"A regularized optimization problem","title":"Solving the problem","text":"We can now choose one of the solvers presented here to solve the problem we defined above. Please refer to other sections of this documentation to make the wisest choice for your particular problem. Depending on the problem structure and on requirements from the user, some solvers are more appropriate than others. The following tries to give a quick overview of what choices one can make.\n\nSuppose for example that we don't want to use a quasi-Newton approach and that we don't have access to the Hessian of f, or that we don't want to incur the cost of computing it.  In this case, the most appropriate solver would be R2. For this example, we also choose a tolerance by specifying the keyword arguments atol and rtol across all solvers.\n\nusing RegularizedOptimization\n \nout = R2(regularized_pb, verbose = 100, atol = 1e-6, rtol = 1e-6)\nprintln(\"R2 converged after $(out.iter) iterations to the solution x = $(out.solution)\")\n\nNow, we can actually use second information on f.  To do so, we are going to use TR, a trust-region solver that can exploit second order information.\n\n\nout = TR(regularized_pb, verbose = 100, atol = 1e-6, rtol = 1e-6)\nprintln(\"TR converged after $(out.iter) iterations to the solution x = $(out.solution)\")\n\nSuppose for some reason we can not compute the Hessian.  In this case, we can try to switch to a quasi-Newton approximation, this can be done
[…]
