### 2025-12-07 47f9a9f
fix wrappedallcos macro for julia pre
