+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"examples/basic/#A-regularized-optimization-problem","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"","category":"section"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"In this tutorial, we will show how to model and solve the nonconvex nonsmooth optimization problem","category":"page"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"  min_x in mathbbR^2 (1 - x_1)^2 + 100(x_2 - x_1^2)^2 + x_1 + x_2","category":"page"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"which can be seen as a ell_1 regularization of the Rosenbrock function.  It can be shown that the solution to the problem is ","category":"page"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"  x^* = beginpmatrix\n  025\n  00575\n  endpmatrix","category":"page"},{"location":"examples/basic/#Modelling-the-problem","page":"A regularized optimization problem","title":"Modelling the problem","text":"","category":"section"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"We first formulate the objective function as the sum of a smooth function f and a nonsmooth regularizer h:","category":"page"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"  (1 - x_1)^2 + 100(x_2 - x_1^2)^2 + x_1 + x_2 = f(x_1 x_2) + h(x_1 x_2)","category":"page"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"where ","category":"page"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"beginalign*\nf(x_1 x_2) = (1 - x_1)^2 + 100(x_2 - x_1^2)^2\nh(x_1 x_2) = x_1\nendalign*","category":"page"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"To model f, we are going to use ADNLPModels.jl. For the nonsmooth regularizer, we use ProximalOperators.jl.  We then wrap the smooth function and the regularizer in a RegularizedNLPModel","category":"page"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularized optimization problem","text":"using ADNLPModels\nusing ProximalOperators\nusing RegularizedProblems\n\n# Model the function\nf_fun = x -> (1 - x[1])^2 + 100*(x[2] - x[1]^2)^2\n\n# Choose a starting point for the optimization process, for the sake of this example, we choose\nx0 = [-1.0, 2.0]\n\n# Get an NLPModel corresponding to the smooth function f\nf_model = ADNLPModel(f_fun, x0, name = \"AD model of f\") \n\n# Get the regularizer from ProximalOperators\nh = NormL1(1.0)\n\n# Wrap into a RegularizedNLPModel\nregularized_pb = RegularizedNLPModel(f_model, h)","category":"page"},{"location":"examples/basic/#Solving-the-problem","page":"A regularized optimization problem","title":"Solving the problem","text":"","category":"section"},{"location":"examples/basic/","page":"A regularized optimization problem","title":"A regularize
[…]
