### 2025-12-08 9fc44c4
Update UUID in Project.toml
The previous UUID was the same as [PosteriorPlots.jl](https://github.com/hendersontrent/PosteriorPlots.jl)
