### 2025-12-09 a5ac6ec
update authors in citation file
Updated citation entry with additional authors.
