+++ CITATION.bib
@@ -1,5 +1,5 @@
-@Misc{baraldi-leconte-orban-regularized-optimization-2024,
-  author = {R. Baraldi and G. Leconte and D. Orban},
+@Misc{baraldi-diouane-gollier-habiboullah-leconte-orban-regularized-optimization-2024,
+  author = {R. Baraldi and and Y. Diouane and M. Gollier and M. L. Habiboullah and G. Leconte and D. Orban},
   title = {{RegularizedOptimization.jl}: Algorithms for Regularized Optimization},
   month = {September},
   howpublished = {\url{https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl}},
