### 2025-12-09 18b22ba
remove typo in citation.bib
