### 2025-12-18 59ba728
move wrappedallocs macro to utils file
