### 2025-12-18 80979c4
add rosenbrock test function
