### 2025-12-18 33d0800
add test solver function
