### 2025-12-18 f4ff437
apply test solver function for R2N
