### 2026-01-03 80070be
add fixmes for lsr1 allocation tests
