### 2026-01-03 1a3e079
add callback test for R2N
