### 2026-01-03 cbf4ab9
mention open PR in fixmes
