### 2026-01-03 ad20f48
mention divide by 0 in fixme
