### 2026-01-03 8dcdd0b
remove test from callback and add in test-solver instead
