### 2026-01-04 b8f941b
Avoid to compute norm of scp  twice in LMSolver, R2DHSolver, and R2Solver
