+++ src/R2DH.jl
@@ -292,6 +292,7 @@ function SolverCore.solve!(
 
   local ξ::T
   local ρk::T = zero(T)
+  local norm_s::T = zero(T)
 
   fk = obj(nlp, xk)
   grad!(nlp, xk, ∇fk)
@@ -328,7 +329,8 @@ function SolverCore.solve!(
   mk(d)::T = φ(d) + ψ(d)::T
 
   spectral_test ? prox!(s, ψ, mν∇fk, ν₁) : iprox!(s, ψ, ∇fk, dkσk)
-  set_solver_specific!(stats, :scp_norm, norm(s))
+  norm_s = norm(s)
+  set_solver_specific!(stats, :scp_norm, norm_s)
 
   mks = mk(s)
 
@@ -380,7 +382,7 @@ function SolverCore.solve!(
           ρk,
           σk,
           norm(xk),
-          norm(s),
+          norm_s,
           (η2 ≤ ρk < Inf) ? '↘' : (ρk < η1 ? '↗' : '='),
         ],
         colsep = 1,
@@ -426,7 +428,8 @@ function SolverCore.solve!(
     m_monotone > 1 && (m_fh_hist[stats.iter % (m_monotone - 1) + 1] = fk + hk)
 
     spectral_test ? prox!(s, ψ, mν∇fk, ν₁) : iprox!(s, ψ, ∇fk, dkσk)
-    set_solver_specific!(stats, :scp_norm, norm(s))
+    norm_s = norm(s)
+    set_solver_specific!(stats, :scp_norm, norm_s)
     mks = mk(s)
 
     ξ = hk - mks + max(1, abs(hk)) * 10 * eps()
@@ -455,7 +458,7 @@ function SolverCore.solve!(
   end
 
   if verbose > 0 && stats.status == :first_order
-    @info log_row(Any[stats.iter, fk, hk, sqrt_ξ_νInv, ρk, σk, norm(xk), norm(s), ""], colsep = 1)
+    @info log_row(Any[stats.iter, fk, hk, sqrt_ξ_νInv, ρk, σk, norm(xk), norm_s, ""], colsep = 1)
     @info "R2DH: terminating with √(ξ/ν) = $(sqrt_ξ_νInv)"
   end
 
