+++ src/R2_alg.jl
@@ -382,6 +382,7 @@ function SolverCore.solve!(
 
   local ξ::T
   local ρk::T
+  local norm_s::T = zero(T)
   σk = max(1 / ν, σmin)
   ν = 1 / σk
   sqrt_ξ_νInv = one(T)
@@ -402,7 +403,8 @@ function SolverCore.solve!(
   mk(d)::T = φk(d) + ψ(d)::T
 
   prox!(s, ψ, mν∇fk, ν)
-  set_solver_specific!(stats, :scp_norm, norm(s))
+  norm_s = norm(s)
+  set_solver_specific!(stats, :scp_norm, norm_s)
   mks = mk(s)
 
   ξ = hk - mks + max(1, abs(hk)) * 10 * eps()
@@ -454,7 +456,7 @@ function SolverCore.solve!(
           ρk,
           σk,
           norm(xk),
-          norm(s),
+          norm_s,
           (η2 ≤ ρk < Inf) ? '↘' : (ρk < η1 ? '↗' : '='),
         ],
         colsep = 1,
@@ -490,7 +492,8 @@ function SolverCore.solve!(
     set_time!(stats, time() - start_time)
 
     prox!(s, ψ, mν∇fk, ν)
-    set_solver_specific!(stats, :scp_norm, norm(s))
+    norm_s = norm(s)
+    set_solver_specific!(stats, :scp_norm, norm_s)
     mks = mk(s)
 
     ξ = hk - mks + max(1, abs(hk)) * 10 * eps()
@@ -519,7 +522,7 @@ function SolverCore.solve!(
   end
 
   if verbose > 0 && stats.status == :first_order
-    @info log_row(Any[stats.iter, fk, hk, sqrt_ξ_νInv, ρk, σk, norm(xk), norm(s), ""], colsep = 1)
+    @info log_row(Any[stats.iter, fk, hk, sqrt_ξ_νInv, ρk, σk, norm(xk), norm_s, ""], colsep = 1)
     @info "R2: terminating with √(ξ/ν) = $(sqrt_ξ_νInv)"
   end
 
