+++ test/runtests.jl
@@ -14,10 +14,12 @@ using ADNLPModels,
 Random.seed!(0)
 const global compound = 1
 const global nz = 10 * compound
-const global options = ROSolverOptions(ν = 1.0, β = 1e16, ϵa = 1e-6, ϵr = 1e-6, verbose = 10)
+const global options =
+  ROSolverOptions(ν = 1.0, β = 1e16, ϵa = 1e-6, ϵr = 1e-6, verbose = 10, neg_tol = 1e-2)
 const global bpdn, bpdn_nls, sol = bpdn_model(compound)
 const global bpdn2, bpdn_nls2, sol2 = bpdn_model(compound, bounds = true)
 const global λ = norm(grad(bpdn, zeros(bpdn.meta.nvar)), Inf) / 10
+const global p_val = 1.2 # for TVp regularization in iR2/iR2N
 
 include("test_AL.jl")
 
@@ -133,6 +135,26 @@ for (mod, mod_name) ∈ (
       end
     end
   end
+end
+
+# iR2/iR2N
+context = ProxTVContext(bpdn.meta.nvar, :lp, p_val, κs = 0.9, λ = λ)
+hp = NormLp(λ, p_val, context)
+@testset "bpdn-iR2N-Lp" begin
+  x0 = zeros(bpdn.meta.nvar)
+  out = iR2N(LBFGSModel(bpdn), hp, options, x0 = x0)
+  @test typeof(out.solution) == typeof(bpdn.meta.x0)
+  @test length(out.solution) == bpdn.meta.nvar
+  @test typeof(out.dual_feas) == eltype(out.solution)
+  @test out.status == :first_order
+end
+@testset "bpdn-iR2-Lp" begin
+  x0 = zeros(bpdn.meta.nvar)
+  out = iR2(bpdn, hp, options, x0 = x0)
+  @test typeof(out.solution) == typeof(bpdn.meta.x0)
+  @test length(out.solution) == bpdn.meta.nvar
+  @test typeof(out.dual_feas) == eltype(out.solution)
+  @test out.status == :first_order
 end
 
 include("test_bounds.jl")
