+++ Project.toml
@@ -1,11 +1,6 @@
 name = "RegularizedOptimization"
 uuid = "20620ad1-4fe4-4467-ae46-fb087718fe7b"
-author = ["Robert Baraldi <rbaraldi@uw.edu>", 
-          "Youssef Diouane <youssef.diouane@polymtl.ca>",
-          "Maxence Gollier <maxence-2.gollier@polymtl.ca>",
-          "Mohamed Laghdaf Habiboullah <mohamed-laghdaf-2.habiboullah@polymtl.ca>",
-          "Geoffroy Leconte <geoffroy.leconte@polymtl.ca>",
-          "Dominique Orban <dominique.orban@gmail.com>"]
+author = ["Robert Baraldi <rbaraldi@uw.edu>", "Youssef Diouane <youssef.diouane@polymtl.ca>", "Maxence Gollier <maxence-2.gollier@polymtl.ca>", "Mohamed Laghdaf Habiboullah <mohamed-laghdaf-2.habiboullah@polymtl.ca>", "Geoffroy Leconte <geoffroy.leconte@polymtl.ca>", "Dominique Orban <dominique.orban@gmail.com>"]
 version = "0.1.0"
 
 [deps]
@@ -33,6 +28,7 @@ NLPModels = "0.19, 0.20, 0.21"
 NLPModelsModifiers = "0.7"
 OptimizationProblems = "0.9.2"
 Percival = "0.7.2"
+ProxTV = "1.2.3"
 ProximalOperators = "0.15"
 RegularizedProblems = "0.1.1"
 ShiftedProximalOperators = "0.2"
