### 2026-01-05 b7a07b1
add opnorm_upper_bound function
