### 2026-01-05 429d036
add opnorm_upper_bound in TR and R2N and add !isnan tests for norm methods
