+++ src/utils.jl
@@ -28,7 +28,7 @@ end
 # ‖Bₖ‖₂ ≤ ‖B₀‖₂ + ∑ᵢ ‖bᵢ‖₂².
 function opnorm_upper_bound(B::LBFGSOperator{T}) where{T} 
   data = B.data
-  bound = data.scaling ? 1/data.scaling_factor : T(1)
+  bound = data.scaling && !iszero(data.scaling_factor) ? 1/abs(data.scaling_factor) : T(1)
   @inbounds for i = 1:data.mem
     bound += norm(data.b[i], 2)^2
   end
@@ -39,7 +39,7 @@ end
 # ‖Bₖ‖₂ ≤ ‖B₀‖₂ + ∑ᵢ |σᵢ|‖aᵢ‖₂².
 function opnorm_upper_bound(B::LSR1Operator{T}) where{T}
   data = B.data
-  bound = data.scaling ? 1/data.scaling_factor : T(1)
+  bound = data.scaling && !iszero(data.scaling_factor) ? 1/abs(data.scaling_factor) : T(1)
   @inbounds for i = 1:data.mem
     if data.as[i] != 0
       bound += norm(data.a[i])^2/abs(data.as[i])
