### 2026-01-06 dbb50c3
add unit tests
