+++ test/runtests.jl
@@ -14,8 +14,7 @@ using ADNLPModels,
 Random.seed!(0)
 const global compound = 1
 const global nz = 10 * compound
-const global options =
-  ROSolverOptions(ν = 1.0, β = 1e16, ϵa = 1e-6, ϵr = 1e-6, verbose = 10, neg_tol = 1e-2)
+const global options = ROSolverOptions(ν = 1.0, β = 1e16, ϵa = 1e-6, ϵr = 1e-6, verbose = 10)
 const global bpdn, bpdn_nls, sol = bpdn_model(compound)
 const global bpdn2, bpdn_nls2, sol2 = bpdn_model(compound, bounds = true)
 const global λ = norm(grad(bpdn, zeros(bpdn.meta.nvar)), Inf) / 10
@@ -138,8 +137,8 @@ for (mod, mod_name) ∈ (
 end
 
 # iR2/iR2N
-context = ProxTVContext(bpdn.meta.nvar, :lp, p_val, κs = 0.9, λ = λ)
-hp = NormLp(λ, p_val, context)
+context = ProxTVContext(bpdn.meta.nvar, :lp, p_val, κs = 0.9, λ = 0.1)
+hp = NormLp(0.1, p_val, context)
 @testset "bpdn-iR2N-Lp" begin
   x0 = zeros(bpdn.meta.nvar)
   out = iR2N(LBFGSModel(bpdn), hp, options, x0 = x0)
