### 2026-01-19 2a3d495
Update installation instructions
