### 2026-01-19 ee643e4
add a contributing section in the README
