### 2026-01-22 83e66ad
Address reviewer comment (#273)
* import SolverCore.reset! explicitely

* fix wrappedallcos macro for julia pre

* bump version to 0.2.0

* revert to v0.1.0 and update author list

* Fix author list formatting

* Add Geoffroy to author list

* Correct author email addresses in Project.toml

Updated author email addresses for Maxence Gollier and Mohamed Laghdaf Habiboullah.

* Update UUID in Project.toml

The previous UUID was the same as [PosteriorPlots.jl](https://github.com/hendersontrent/PosteriorPlots.jl)

* update authors in citation file

Updated citation entry with additional authors.


[…]
