+++ Project.toml
@@ -1,6 +1,11 @@
 name = "RegularizedOptimization"
-uuid = "196f2941-2d58-45ba-9f13-43a2532b2fa8"
-author = ["Robert Baraldi <rbaraldi@uw.edu> and Dominique Orban <dominique.orban@gmail.com>"]
+uuid = "20620ad1-4fe4-4467-ae46-fb087718fe7b"
+author = ["Robert Baraldi <rbaraldi@uw.edu>", 
+          "Youssef Diouane <youssef.diouane@polymtl.ca>",
+          "Maxence Gollier <maxence-2.gollier@polymtl.ca>",
+          "Mohamed Laghdaf Habiboullah <mohamed-laghdaf-2.habiboullah@polymtl.ca>",
+          "Geoffroy Leconte <geoffroy.leconte@polymtl.ca>",
+          "Dominique Orban <dominique.orban@gmail.com>"]
 version = "0.1.0"
 
 [deps]
