+++ docs/make.jl
@@ -27,4 +27,5 @@ makedocs(
 deploydocs(
   repo = "github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl.git",
   push_preview = true,
+  devbranch = "main",
 )
