### 2026-01-26 c526521
add optional multipliers
