### 2026-01-26 0871eee
add compute_obj and compute_grad options
