### 2026-01-26 7540372
add compute_obj and compute_grad options
