### 2026-01-27 #9344 @MohamedLaghdafHABIBOULLAH
Thank you for your reviews, @rafaelbailo and @pierre-borie, and thank you @fabian-s for agreeing to edit this submission.
Here are our responses:

* Installation: We updated the README now that the package has been registered.

* Unit testing: Proximal splitting, proximal gradient descent, and FISTA are no longer part of the package, since it is dedicated to regularized and trust-region methods. These components were already removed in this [PR](https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl/pull/207) prior to submitting to JOSS.

* Examples: We ran the example on macOS with Julia 1.11.7, we added this information to the paper in the revision and documented the exact steps for reproducibility in this [issue](https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl/issues/259). Note that results may vary across operating systems and Julia versions.

* Community guidelines: We added a link to the JSO contribution guidelines in the README: https://jso.dev/contributing/.

Best regards,
@d1Lab @MaxenceGollier @MohamedLaghdafHABIBOULLAH @dpo
on behalf of the authors
