### 2026-01-27 8871987
add matrix and arpack tests
