### 2026-01-27 f385d85
simplify condition for arpack
