### 2026-01-29 4c4773f
change the callback signature to use reg_nlp instead of nlp
