+++ src/LM_alg.jl
@@ -327,7 +327,7 @@ function SolverCore.solve!(
     ),
   )
 
-  callback(nls, solver, stats)
+  callback(reg_nls, solver, stats)
 
   done = stats.status != :unknown
 
@@ -463,7 +463,7 @@ function SolverCore.solve!(
       ),
     )
 
-    callback(nls, solver, stats)
+    callback(reg_nls, solver, stats)
 
     done = stats.status != :unknown
   end
