### 2026-01-29 bafbfaf
change the callback signature to use reg_nlp instead of nlp
