### 2026-02-02 #9344 @fabian-s
@pierre-borie based on your thumbs-up response, I assume that you now assess that "functional claims of the software [have] been confirmed", your last missing checklist item?
