### 2026-02-02 JuliaSmoothOptimizers/RegularizedOptimization.jl#284 @fabian-s
JOSS paper: Typos/Phrasing
Please consider adopting some of these suggestions:

Typos/Errors:

- Line 42: "of the form \eqref{eq:nlp} inexactly, in which $f$, $h$, and $c$
  are replaced with appropriate models about the current iterate."  ::   "models
  about" should be "models around" (standard optimization terminology: model
  around the current iterate).
- Line 48: "a nonsmooth penalty approach, described in
  [@diouane-gollier-orban-2024] is currently being developed"  ::   Missing comma:
  should be "described in [@diouane-gollier-orban-2024]**,** is currently being
  developed".
- Line 86: "They can also be paired into a *Regularized Nonlinear
  Least-Squares Model* if $f(x) = \tfrac{1}{2} \|F(x)\|^2$ for some residual $F:
   \mathbb{R}^n \to \mathbb{R}^m$, in the case of the **LM** and **LMTR**
  solvers."  ::   Awkward phrasing. The "in the case of" clause dangles. Suggest:
  "They can also be paired into a Regularized Nonlinear Least-Squares Model,
  used by the LM and LMTR solvers, if $f(x) = \tfrac{1}{2} |F(x)|^2$ for some
  residual $F: \mathbb{R}^n \to \mathbb{R}^m$."
- Line 97: "[RegularizedOptimization.jl](https://github.com/JuliaSmoothOptimiz
  ers/RegularizedOptimization.jl), methods such as **R2N** and **TR** methods
  support"  ::   "methods" appears twice: "methods such as **R2N** and **TR**
  methods". Remove one: "methods such as R2N and TR support".
- Line 97: "In contrast with"  ::   conventionally "In contrast to" or "By
  contrast with".
- Line 108: "showing how to define and solve the problem, and perform a solve
  followed by a re-solve"  ::   Redundant: "solve the problem, and perform a solve".
   Suggest: "showing how to define the problem and perform a solve followed by a
   re-solve".
- Line 126: "The table reports the convergence status of each solver, the
  number of evaluations of $f$, the number of evaluations of $\nabla f$, the
  number of proximal operator evaluations, the elapsed time and the final
  objective value."  ::   Missing Oxford comma before "and the final objective
  value" (used elsewhere in the paper, so should be consistent).

Phrasing suggestions:

- Line 44: "All solvers are implemented in place, so re-solves incur no
  allocations."  ::   "implemented in place" is slightly ambiguous (could mean
  in-place operations vs. implemented locally). Consider "All solvers operate in
   place" or "All solvers are allocation-free, so re-solves incur no additional
  allocations."
- Line 139: "Ongoing research aims to reduce the number of proximal
  evaluations."  ::   Feels abrupt as a closing sentence for the results section.
  Could add a brief forward-looking clause with a pointer to this
  research?

https://github.com/openjournals/joss-reviews/issues/9344
