### 2026-02-02 #9344 @fabian-s
@MohamedLaghdafHABIBOULLAH please see https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl/issues/284

Once that is done, please prepare the archival release as detailed below so I can recommend this for acceptance.
