### 2026-02-02 127c4ac
compute hess_op with respect to regularizedNLPModel
