### 2026-02-02 fdbf4b7
compute hess_op with respect to regularizedNLPModel
