+++ Project.toml
@@ -33,7 +33,7 @@ NLPModelsModifiers = "0.7"
 OptimizationProblems = "0.9.2"
 Percival = "0.7.2"
 ProximalOperators = "0.15"
-RegularizedProblems = "0.1.1"
+RegularizedProblems = "0.1.3"
 ShiftedProximalOperators = "0.2"
 SolverCore = "0.3.0"
 julia = "^1.6.0"
