### 2026-02-04 18bf6c1
use linearOperators 2.12 and new op_upper_bound feature for SR1 and BFGS approximations
