### 2026-02-04 b630a18
Bump to 0.2.0
