+++ Project.toml
@@ -6,7 +6,7 @@ author = ["Robert Baraldi <rbaraldi@uw.edu>",
           "Mohamed Laghdaf Habiboullah <mohamed-laghdaf-2.habiboullah@polymtl.ca>",
           "Geoffroy Leconte <geoffroy.leconte@polymtl.ca>",
           "Dominique Orban <dominique.orban@gmail.com>"]
-version = "0.1.0"
+version = "0.2.0"
 
 [deps]
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
