+++ paper/paper.bib
@@ -159,3 +159,14 @@
   publisher     = {Springer},
   doi           = {10.1007/BF01581204}
 }
+
+@techreport{allaire-le-digabel-orban-2025,
+  title       = {An inexact modified quasi-Newton method for nonsmooth regularized optimization},
+  author      = {Nathan Allaire and S{\'e}bastien Le Digabel and Dominique Orban},
+  institution = {GERAD},
+  type        = {Cahier},
+  number      = {G-2025-73},
+  year        = {2025},
+  address     = {Montr{\'e}al, Canada},
+  url         = {10.13140/RG.2.2.32728.97288}
+}
