### 2026-02-04 61189af
Update paper/paper.md
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
