### 2026-02-04 ffcfaea
Update paper/paper.bib
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
