+++ paper/paper.bib
@@ -168,5 +168,5 @@
   number      = {G-2025-73},
   year        = {2025},
   address     = {Montr{\'e}al, Canada},
-  url         = {10.13140/RG.2.2.32728.97288}
+  doi         = {10.13140/RG.2.2.32728.97288}
 }
