### 2026-02-05 #9344 @fabian-s
@MohamedLaghdafHABIBOULLAH thanks, what I need from you now so I can finalize this is:

- **Make a release of the software** with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
- Archive the release on Zenodo/figshare/etc and **post the DOI here.**
