### 2026-02-05 47a1acf
Update RegularizedOptimization dependency in Project.toml to new UUID and set compatibility version
