### 2026-02-09 #9344 @MaxenceGollier
@fabian-s, sorry, we had difficulties with zenodo.

-  I double checked authors and affiliations
- We released the [`v0.2.0`](https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl/releases/tag/v0.2.0) with the latest changes from the review.
- We  archived the release on Zenodo, the DOI is [10.5281/zenodo.18486418](https://doi.org/10.5281/zenodo.18486418).
- I double checked the title and author list in the archive.
- I double checked the license listed for the archive.
