### 2026-02-09 #9344 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
