### 2026-02-09 #9344 @dpo
@fabian-s While updating the Zenodo record of this package, we realized that several other authors contributed significant amounts of code in the past and deserve to appear in the list of authors. Would it be acceptable to make such change at this point?

For reference here is the link to the contributors: https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl/graphs/contributors

We're proposing to add geoffroyleconte, rjbaraldi and aldma to recognize their contributions formally.
