### 2026-02-09 #9344 @danielskatz
👋 @dpo & @MaxenceGollier - we can add authors to the paper and the zenodo archive, assuming the reviewers agree this is ok.  If not, you can add them to the zenodo archive in any case.
👋 @rafaelbailo, @pierre-borie - Can we get your opinion on adding these three authors?
