### 2026-02-09 #9344 @pierre-borie
@danielskatz No problem on my end ! I see no issues with adding these authors.
