### 2026-02-11 #9344 @danielskatz
👋 @dpo & @MaxenceGollier - please go ahead and update the authors.  Let me know when you have done so, and I'll then proofread the paper and move it to the next step.
