### 2026-02-12 11be3dd
Update authors and affiliations in paper.md
