### 2026-02-13 9a4ba1a
minor changes in paper.md
