+++ paper/paper.bib
@@ -161,7 +161,7 @@
 }
 
 @techreport{allaire-le-digabel-orban-2025,
-  title       = {An inexact modified quasi-Newton method for nonsmooth regularized optimization},
+  title       = {An inexact modified quasi-{N}ewton method for nonsmooth regularized optimization},
   author      = {Nathan Allaire and S{\'e}bastien Le Digabel and Dominique Orban},
   institution = {GERAD},
   type        = {Cahier},
