### 2026-02-13 #9344 @danielskatz
@MohamedLaghdafHABIBOULLAH - and also, I have a few small changes that I suggest in https://github.com/JuliaSmoothOptimizers/RegularizedOptimization.jl/pull/291 - it might be easier to merge this and then add the missing affiliation countries. Of course, if you disagree with anything, let me know.
