### 2026-02-13 f93b452
Update affiliations in paper.md with full country names
