### 2026-02-13 33a583d
add sparse hessian
