### 2026-02-13 6ae6e2e
add _reset_power_norm function
