+++ src/R2N.jl
@@ -225,6 +225,7 @@ function SolverCore.solve!(
   sub_kwargs::NamedTuple = NamedTuple(),
 ) where {T, V, G}
   reset!(stats)
+  _reset_power_method!(solver.v0)
 
   # Retrieve workspace
   selected = reg_nlp.selected
