### 2026-02-13 c194e13
add LinearOperators.reset! to reset! quasi-Newton approximations
