+++ src/R2N.jl
@@ -226,6 +226,7 @@ function SolverCore.solve!(
 ) where {T, V, G}
   reset!(stats)
   _reset_power_method!(solver.v0)
+  LinearOperators.reset!(reg_nlp.model)
 
   # Retrieve workspace
   selected = reg_nlp.selected
