### 2026-02-13 302a7b9
fix _reset_power_method to correct output
