### 2026-02-13 73836e7
implement SolverCore.reset!(solver) functions
