### 2026-02-13 3c898ec
add condition on B before resetting
