+++ src/R2N.jl
@@ -98,7 +98,8 @@ end
 
 function SolverCore.reset!(solver::R2NSolver)
   _reset_power_method!(solver.v0)
-  LinearOperators.reset!(solver.subpb.model.B)
+  B = solver.subpb.model.B
+  isa(B, AbstractLinearOperator) && LinearOperators.reset!(B)
 end
 
 SolverCore.reset!(solver::R2NSolver, model) = SolverCore.reset!(solver)
