### 2026-02-15 #9344 @danielskatz
Congratulations to @MaxenceGollier (Maxence Gollier) and co-authors on your JOSS publication!!

JOSS fully depends on volunteers - we couldn't do this without you
