### 2026-02-18 #9344 @MaxenceGollier
Thanks @danielskatz and @fabian-s for handling this submission. Many thanks as well to @pierre-borie @rafaelbailo for taking the time to review our paper. Thank you all for your helpful and constructive comments! 👍 🥳 👍
